Revert "Fix small shifts (use double)"
This reverts commit 9798a3b55cf7ff5ad41657226f1add5a88a9290c.

+++ OsmAnd-java/src/net/osmand/RenderingContext.java
@@ -29,13 +29,13 @@ public class RenderingContext {
 	}
 	
 
-	public double leftX;
-	public double topY;
+	public float leftX;
+	public float topY;
 	public int width;
 	public int height;
 
 	public int zoom;
-	public double tileDivisor;
+	public float tileDivisor;
 	public float rotate;
 
 	// debug purpose
